requirements.txt
# LangChain Custom Chatbot with Memory - Dependencies
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.12.0
python-dotenv==1.0.0
streamlit==1.29.0
chromadb==0.4.22
tiktoken==0.5.2
pydantic==2.5.3
typing-extensions==4.9.0 